Попытка убрать конфликты для добавления ветки по удалению фильмов и пользователей

+++ src/main/java/ru/yandex/practicum/filmorate/controller/UserController.java
@@ -1,20 +1,10 @@
 package ru.yandex.practicum.filmorate.controller;
 
 import jakarta.validation.Valid;
-import jakarta.validation.constraints.Positive;
 import lombok.AllArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import org.springframework.http.HttpStatus;
-import org.springframework.web.bind.annotation.DeleteMapping;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.PutMapping;
-import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.ResponseStatus;
-import org.springframework.web.bind.annotation.RestController;
-import ru.yandex.practicum.filmorate.model.Film;
+import org.springframework.web.bind.annotation.*;
+import ru.yandex.practicum.filmorate.exception.NotFoundException;
 import ru.yandex.practicum.filmorate.model.User;
 import ru.yandex.practicum.filmorate.service.UserService;
 
@@ -25,7 +15,7 @@ import java.util.List;
 @RestController
 @AllArgsConstructor
 @RequestMapping("/users")
-public class UserController{
+public class UserController {
     private final UserService userService;
 
 
@@ -73,6 +63,10 @@ public class UserController{
 
     @DeleteMapping("/{id}")
     public void deleteUser(@PathVariable Long id) {
+        if (id > 0) {
             userService.deleteUser(id);
+        } else {
+            throw new NotFoundException("Пользователь с id " + id + " не существует");
+        }
     }
 }
